@@ -1,6 +1,6 @@
 {
   "name": "nypl-discovery",
-  "version": "1.2.9",
+  "version": "1.4.1",
   "description": "Isomorphic React App for NYPL Research Catalog.",
   "main": "index.js",
   "scripts": {
@@ -56,11 +56,12 @@
     "chai": "3.5.0",
     "create-react-class": "15.5.3",
     "esdoc": "0.4.5",
-    "eslint": "4.9.0",
+    "eslint": "4.15.0",
     "eslint-config-airbnb": "16.1.0",
     "eslint-loader": "1.3.0",
-    "eslint-plugin-jsx-a11y": "6.0.2",
-    "eslint-plugin-react": "7.4.0",
+    "eslint-plugin-import": "2.8.0",
+    "eslint-plugin-jsx-a11y": "6.0.3",
+    "eslint-plugin-react": "7.5.1",
     "jsdom": "9.0.0",
     "mocha": "4.0.1",
     "nightwatch": "0.9.9",
@@ -70,8 +71,8 @@
     "webpack-dev-server": "2.7.1"
   },
   "dependencies": {
-    "@nypl/design-toolkit": "0.1.27",
-    "@nypl/dgx-header-component": "2.4.5",
+    "@nypl/design-toolkit": "0.1.35",
+    "@nypl/dgx-header-component": "2.4.7",
     "@nypl/dgx-react-footer": "0.4.1",
     "@nypl/dgx-svg-icons": "0.2.5",
     "@nypl/nypl-data-api-client": "0.2.5",
@@ -94,6 +95,7 @@
     "dgx-react-ga": "git+https://git@bitbucket.org/NYPL/dgx-react-ga.git#master",
     "ejs": "2.5.5",
     "enzyme": "2.9.1",
+    "eslint-loader": "1.3.0",
     "express": "4.13.1",
     "extract-text-webpack-plugin": "2.1.2",
     "focus-trap-react": "3.0.3",
