@@ -1,7 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router';
-import Actions from '../../actions/Actions.js';
 import {
   isArray as _isArray,
   isEmpty as _isEmpty,
@@ -10,12 +9,13 @@ import {
   every as _every,
 } from 'underscore';
 
+import Actions from '../../actions/Actions';
 import {
   ajaxCall,
   trackDiscovery,
 } from '../../utils/utils';
 import DefinitionList from './DefinitionList';
-import appConfig from '../../../../appConfig.js';
+import appConfig from '../../../../appConfig';
 
 class BibDetails extends React.Component {
   /*
@@ -35,13 +35,15 @@ class BibDetails extends React.Component {
       return null;
     }
 
-    items.forEach(item => {
+    items.forEach((item) => {
       const ownerObj = item.owner && item.owner.length ? item.owner[0].prefLabel : undefined;
 
       ownerArr.push(ownerObj);
     });
 
-    if (_every(ownerArr, (o) => (o === ownerArr[0]))) {
+    // From above, check to see if all the owners are the same, and if so, proceed if the owner
+    // is either Princeton or Columbia.
+    if (_every(ownerArr, o => (o === ownerArr[0]))) {
       if ((ownerArr[0] === 'Princeton University Library') ||
         (ownerArr[0] === 'Columbia University Libraries')) {
         owner = ownerArr[0];
@@ -52,7 +54,23 @@ class BibDetails extends React.Component {
   }
 
   /*
-   * getDefinitionObject(bibValues, fieldValue, fieldLinkable)
+   * Return note array or null.
+   * @param {object} bib
+   * @return {null|array}
+   */
+  getNote(bib) {
+    const note = bib.note;
+    let notes = note && note.length ? note : null;
+
+    if (!notes) {
+      return null;
+    }
+
+    return notes;
+  }
+
+  /*
+   * getDefinitionObject(bibValues, fieldValue, fieldLinkable, fieldSelfLinkable, fieldLabel)
    * Gets a list, or one value, of data to display for a field from the API, where
    * the data is an object in the array.
    * @param {array} bibValues
@@ -62,6 +80,7 @@ class BibDetails extends React.Component {
    * @param {string} fieldLabel
    */
   getDefinitionObject(bibValues, fieldValue, fieldLinkable, fieldSelfLinkable, fieldLabel) {
+    // If there's only one value, we just want that value and not a list.
     if (bibValues.length === 1) {
       const bibValue = bibValues[0];
       const url = `filters[${fieldValue}]=${bibValue['@id']}`;
@@ -97,12 +116,12 @@ class BibDetails extends React.Component {
           bibValues.map((value, i) => {
             const url = `filters[${fieldValue}]=${value['@id']}`;
             let itemValue = fieldLinkable ?
-              <Link
+              (<Link
                 onClick={e => this.newSearch(e, url, fieldValue, value['@id'], fieldLabel)}
                 to={`${appConfig.baseUrl}/search?${url}`}
               >
                 {value.prefLabel}
-              </Link>
+              </Link>)
               : <span>{value.prefLabel}</span>;
             if (fieldSelfLinkable) {
               itemValue =
@@ -131,7 +150,7 @@ class BibDetails extends React.Component {
     let val = '';
 
     if (bibValues.length) {
-      bibValues.forEach(value => {
+      bibValues.forEach((value) => {
         if (value.indexOf(`${fieldIdentifier}:`) !== -1) {
           val = value.substring(fieldIdentifier.length + 1);
         }
@@ -146,16 +165,19 @@ class BibDetails extends React.Component {
   }
 
   /*
-   * getDefinition(bibValues, fieldValue, fieldLinkable, fieldIdentifier)
+   * getDefinition(bibValues, fieldValue, fieldLinkable, fieldIdentifier,
+   * fieldSelfLinkable, fieldLabel)
    * Gets a list, or one value, of data to display for a field from the API.
    * @param {array} bibValues
    * @param {string} fieldValue
    * @param {boolean} fieldLinkable
    * @param {string} fieldIdentifier
    * @param {string} fieldSelfLinkable
+   * @param {string} fieldLabel
    */
   getDefinition(
-    bibValues, fieldValue, fieldLinkable, fieldIdentifier, fieldSelfLinkable, fieldLabel
+    bibValues, fieldValue, fieldLinkable, fieldIdentifier,
+    fieldSelfLinkable, fieldLabel,
   ) {
     if (fieldValue === 'identifier') {
       return this.getIdentifiers(bibValues, fieldIdentifier);
@@ -195,17 +217,17 @@ class BibDetails extends React.Component {
         {
           bibValues.map((value, i) => {
             const url = `filters[${fieldValue}]=${value}`;
-            let itemValue = fieldLinkable ?
+            let itemValue = fieldLinkable ? (
               <Link
                 onClick={e => this.newSearch(e, url, fieldValue, value, fieldLabel)}
                 to={`${appConfig.baseUrl}/search?${url}`}
               >
                 {value}
-              </Link>
+              </Link>)
               : <span>{value}</span>;
             if (fieldSelfLinkable) {
-              itemValue =
-                (<a
+              itemValue = (
+                <a
                   href={value.url}
                   onClick={() => trackDiscovery('Bib fields', `${fieldLabel} - ${value.prefLabel}`)}
                 >
@@ -221,33 +243,6 @@ class BibDetails extends React.Component {
   }
 
   /**
-   * getPublication(bib)
-   * Get an object with publisher detail information.
-   * @param {object} bib
-   * @return {object}
-   */
-  getPublication(bib) {
-    const fields = ['placeOfPublication', 'publisherLiteral', 'createdString'];
-    let publicationInfo = '';
-
-    fields.forEach(field => {
-      const fieldValue = bib[field];
-      if (fieldValue) {
-        publicationInfo += `${fieldValue} `;
-      }
-    });
-
-    if (!publicationInfo) {
-      return null;
-    }
-
-    return {
-      term: 'Publication',
-      definition: <span>{publicationInfo}</span>,
-    };
-  }
-
-  /**
    * getDisplayFields(bib)
    * Get an array of definition term/values.
    * @param {object} bib
@@ -258,7 +253,6 @@ class BibDetails extends React.Component {
     // component rather than from the bib field properties.
     const fields = this.props.fields;
     const fieldsToRender = [];
-    const publicationInfo = this.getPublication(bib);
 
     fields.forEach((field) => {
       const fieldLabel = field.label;
@@ -279,12 +273,14 @@ class BibDetails extends React.Component {
             term: fieldLabel,
             definition:
               this.getDefinitionObject(
-                bibValues, fieldValue, fieldLinkable, fieldSelfLinkable, fieldLabel
+                bibValues, fieldValue, fieldLinkable, fieldSelfLinkable,
+                fieldLabel,
               ),
           });
         } else {
           const definition = this.getDefinition(
-            bibValues, fieldValue, fieldLinkable, fieldIdentifier, fieldSelfLinkable, fieldLabel
+            bibValues, fieldValue, fieldLinkable, fieldIdentifier,
+            fieldSelfLinkable, fieldLabel,
           );
           if (definition) {
             fieldsToRender.push({
@@ -293,11 +289,6 @@ class BibDetails extends React.Component {
             });
           }
         }
-      }
-
-      // This is made up of three different bib property values so it's special.
-      if (fieldLabel === 'Publication' && publicationInfo) {
-        fieldsToRender.push(publicationInfo);
       }
 
       // The Owner is complicated too.
@@ -311,14 +302,55 @@ class BibDetails extends React.Component {
         }
       }
 
+      // Note field rendering as array of objects instead of an array of strings.
+      // Parse the original and new note format.
+      // Original format: ['string1', 'string2']
+      // 2018 format:
+      //    [{'noteType': 'string',
+      //     'prefLabel': 'string',
+      //     '@type': 'bf:Note'},
+      //    {...}]
+      if (fieldLabel === 'Notes') {
+        const note = this.getNote(this.props.bib);
+        // Make sure we have at least one note
+        if (note && Array.isArray(note)) {
+          // Group notes by noteType:
+          const notesGroupedByNoteType = note
+            // Make sure all notes are blanknodes:
+            .filter(n => (typeof n) === 'object')
+            .reduce((groups, n) => {
+              if (!groups[n.noteType]) groups[n.noteType] = [];
+              groups[n.noteType].push(n);
+              return groups;
+            }, {});
+
+          // For each group of notes, add a fieldToRender:
+          Object.keys(notesGroupedByNoteType).forEach((noteType) => {
+            const notesList = (
+              <ul>
+                {
+                  notesGroupedByNoteType[noteType].map((n, i) => (
+                    <li key={i.toString()}>{n.prefLabel}</li>
+                  ))
+                }
+              </ul>
+            );
+            fieldsToRender.push({
+              term: noteType,
+              definition: notesList,
+            });
+          });
+        }
+      }
+
       if (fieldLabel === 'Electronic Resource' && this.props.electronicResources.length) {
         const electronicResources = this.props.electronicResources;
         let electronicElem;
 
         if (electronicResources.length === 1) {
           const electronicItem = electronicResources[0];
-          electronicElem =
-            (<a
+          electronicElem = (
+            <a
               href={electronicItem.url}
               target="_blank"
               onClick={() =>
@@ -359,6 +391,42 @@ class BibDetails extends React.Component {
     return fieldsToRender;
   }
 
+  /*
+   * Display for single and multivalued object arrays.
+   * @param {array} note
+   * @return {string}
+   */
+  noteObjectDisplay(note) {
+    let display;
+    if (note.length === 1) {
+      display = (
+        <div>
+          <h4>{note[0].noteType}</h4>
+          <p>{note[0].prefLabel}</p>
+        </div>
+      );
+    } else {
+      display = (
+        <ul>
+          {
+            note.map((n, i) => (
+              <li key={i.toString()}>
+                <h4>
+                  {n.noteType}
+                </h4>
+                <p>
+                  {n.prefLabel}
+                </p>
+              </li>
+            ))
+          }
+        </ul>
+      );
+    }
+
+    return display;
+  }
+
   newSearch(e, query, field, value, label) {
     e.preventDefault();
 
@@ -369,29 +437,29 @@ class BibDetails extends React.Component {
       let index = 0;
 
       if (response.data.facet) {
-        // Find the index where the field exists in the list of facets from the API
-        index = _findIndex(response.data.facets.itemListElement, { field });
+        // Find the index where the field exists in the list of filters from the API
+        index = _findIndex(response.data.filters.itemListElement, { field });
       }
 
-      // If the index exists, try to find the facet value from the API
-      if (response.data.facets && response.data.facets.itemListElement
-        && response.data.facets.itemListElement[index]) {
-        const facet = _findWhere(response.data.facets.itemListElement[index].values, { value });
+      // If the index exists, try to find the filter value from the API
+      if (response.data.filters && response.data.filters.itemListElement
+        && response.data.filters.itemListElement[index]) {
+        const filter = _findWhere(response.data.filters.itemListElement[index].values, { value });
 
-        // The API may return a list of facets in the selected field, but the wanted
-        // facet may still not appear. If that's the case, return the clicked facet value.
-        Actions.updateSelectedFacets({
+        // The API may return a list of filters in the selected field, but the wanted
+        // filter may still not appear. If that's the case, return the clicked filter value.
+        Actions.updateSelectedFilters({
           [field]: [{
-            id: facet ? facet.value : value,
-            value: facet ? (facet.label || facet.value) : value,
+            value: filter ? filter.value : value,
+            label: filter ? (filter.label || filter.value) : value,
           }],
         });
       } else {
         // Otherwise, the field wasn't found in the API. Returning this highlights the
-        // facet in the selected facet region, but not in the facet sidebar.
-        Actions.updateSelectedFacets({
+        // filter in the selected filter region, but not in the filter sidebar.
+        Actions.updateSelectedFilters({
           [field]: [{
-            id: value,
+            label: value,
             value,
           }],
         });
@@ -404,7 +472,7 @@ class BibDetails extends React.Component {
       Actions.updatePage('1');
       setTimeout(
         () => { this.props.updateIsLoadingState(false); },
-        500
+        500,
       );
       this.context.router.push(`${appConfig.baseUrl}/search?${query}`);
     });
