@@ -1,11 +1,13 @@
+/* globals document */
 import React from 'react';
 import PropTypes from 'prop-types';
 import DocumentTitle from 'react-document-title';
 import { every as _every } from 'underscore';
+import { LeftWedgeIcon } from '@nypl/dgx-svg-icons';
 
 import Breadcrumbs from '../Breadcrumbs/Breadcrumbs';
 import ItemHoldings from '../Item/ItemHoldings';
-import LoadingLayer from '../LoadingLayer/LoadingLayer.jsx';
+import LoadingLayer from '../LoadingLayer/LoadingLayer';
 import BibDetails from './BibDetails';
 import LibraryItem from '../../utils/item';
 import BackLink from './BackLink';
@@ -30,6 +32,10 @@ class BibPage extends React.Component {
     this.updateIsLoadingState = this.updateIsLoadingState.bind(this);
   }
 
+  componentDidMount() {
+    document.getElementById('bib-title').focus();
+  }
+
   updateIsLoadingState(status) {
     this.setState({ isLoading: status });
   }
@@ -40,7 +46,7 @@ class BibPage extends React.Component {
     const bibId = bib && bib['@id'] ? bib['@id'].substring(4) : '';
     const title = bib.title && bib.title.length ? bib.title[0] : '';
     const items = LibraryItem.getItems(bib);
-    const isElectronicResources = _every(items, (i) => i.isElectronicResource);
+    const isElectronicResources = _every(items, i => i.isElectronicResource);
     // Related to removing MarcRecord because the webpack MarcRecord is not working. Sep/28/2017
     // const isNYPLReCAP = LibraryItem.isNYPLReCAP(bib['@id']);
     // const bNumber = bib && bib.idBnum ? bib.idBnum : '';
@@ -64,14 +70,19 @@ class BibPage extends React.Component {
     ];
 
     const bottomFields = [
-      { label: 'Publication', value: 'React Component' },
+      { label: 'Publication', value: 'publicationStatement' },
+      { label: 'Publication Date', value: 'serialPublicationDates' },
       { label: 'Electronic Resource', value: 'React Component' },
       { label: 'Description', value: 'extent' },
+      { label: 'Series Statement', value: 'seriesStatement' },
+      { label: 'Uniform Title', value: 'uniformTitle' },
+      { label: 'Alternative Title', value: 'titleAlt' },
+      { label: 'Former Title', value: 'formerTitle' },
       { label: 'Subject', value: 'subjectLiteral', linkable: true },
-      { label: 'Genre/Form', value: 'materialType' },
-      { label: 'Notes', value: '' },
+      { label: 'Genre/Form', value: 'genreForm' },
+      { label: 'Notes', value: 'React Component' },
       { label: 'Additional Resources', value: 'supplementaryContent', selfLinkable: true },
-      { label: 'Contents', value: 'note' },
+      { label: 'Contents', value: 'tableOfContents' },
       { label: 'Bibliography', value: '' },
       { label: 'ISBN', value: 'identifier', identifier: 'urn:isbn' },
       { label: 'ISSN', value: 'identifier', identifier: 'urn:issn' },
@@ -81,7 +92,7 @@ class BibPage extends React.Component {
       { label: 'Owning Institutions', value: '' },
     ];
 
-    const itemHoldings = items.length && !isElectronicResources ?
+    const itemHoldings = items.length && !isElectronicResources ? (
       <ItemHoldings
         shortenItems={shortenItems}
         items={items}
@@ -89,7 +100,7 @@ class BibPage extends React.Component {
         itemPage={itemPage}
         searchKeywords={this.props.searchKeywords}
         updateIsLoadingState={this.updateIsLoadingState}
-      /> : null;
+      />) : null;
     // Related to removing MarcRecord because the webpack MarcRecord is not working. Sep/28/2017
     // const marcRecord = isNYPLReCAP ? <MarcRecord bNumber={bNumber[0]} /> : null;
 
@@ -107,28 +118,24 @@ class BibPage extends React.Component {
     return (
       <DocumentTitle title="Item Details | Shared Collection Catalog | NYPL">
         <main className="main-page">
-          <LoadingLayer status={this.state.isLoading} title="Searching" />
+          <LoadingLayer
+            status={this.state.isLoading}
+            title="Searching"
+          />
           <div className="nypl-page-header">
             <div className="nypl-full-width-wrapper">
               <div className="nypl-row">
                 <div className="nypl-column-three-quarters">
                   <Breadcrumbs type="bib" query={searchURL} />
-                  <h1>Item Details</h1>
+                  <h1 id="bib-title" tabIndex="0">Item Details</h1>
                   <h2>{title}</h2>
                   {
                     this.props.searchKeywords && (
                       <div className="nypl-row search-control">
-                        <svg
-                          xmlns="http://www.w3.org/2000/svg"
-                          width="25" height="42"
-                          viewBox="0 0 25 42"
+                        <LeftWedgeIcon
                           preserveAspectRatio="xMidYMid meet"
-                          aria-labelledby="left-arrow"
-                          aria-hidden="true"
-                        >
-                          <title id="left-arrow">Back to Results</title>
-                          <polygon points="21.172 42.344 0 21.172 21.172 0 25.112 3.939 7.88 21.172 25.112 38.404 21.172 42.344"></polygon>
-                        </svg>
+                          title="Back to Results"
+                        />
                         <BackLink
                           searchURL={searchURL}
                           searchKeywords={this.props.searchKeywords}
@@ -182,15 +189,10 @@ class BibPage extends React.Component {
 }
 
 BibPage.propTypes = {
-  item: PropTypes.object,
   searchKeywords: PropTypes.string,
   location: PropTypes.object,
-  selectedFacets: PropTypes.object,
   bib: PropTypes.object,
-  field: PropTypes.string,
   isLoading: PropTypes.bool,
-  sortBy: PropTypes.string,
-  page: PropTypes.string,
 };
 
 export default BibPage;
